Restructure Feed Navigator tabs so that Camera and Chat (left and right swipes) are only active on Feed's root page;
