+++ src/navigation/TabNavigator.js
@@ -1,7 +1,6 @@
 import React from 'react'
 import { TouchableOpacity } from 'react-native'
 import { createBottomTabNavigator } from '@react-navigation/bottom-tabs'
-import { createMaterialTopTabNavigator } from '@react-navigation/material-top-tabs'
 import { withTheme } from 'react-native-paper'
 
 import * as navigationOptions from 'navigation/options'
@@ -10,7 +9,6 @@ import * as navigationActions from 'navigation/actions'
 import FeedNavigator from 'navigation/Feed'
 import SearchNavigator from 'navigation/Search'
 import CameraNavigator from 'navigation/Camera'
-import ChatNavigator from 'navigation/Chat'
 import DatingNavigator from 'navigation/Dating'
 import ProfileNavigator from 'navigation/Profile'
 
@@ -19,35 +17,6 @@ import SearchIcon from 'assets/svg/footer/Search'
 import CreateIcon from 'assets/svg/footer/Create'
 import HeartIcon from 'assets/svg/footer/Heart'
 import UserIcon from 'assets/svg/footer/User'
-
-const HomeNavigator = ({ theme }) => {
-  const Tab = createMaterialTopTabNavigator()
-
-  const tabNavigatorProps = {
-    initialRouteName: 'Feed',
-    tabBar: () => null,
-    sceneContainerStyle: {
-      backgroundColor: 'transparent',
-    },
-  }
-
-  return (
-    <Tab.Navigator {...tabNavigatorProps}>
-      <Tab.Screen
-        name="Camera"
-        component={CameraNavigator}
-      />
-      <Tab.Screen
-        name="Feed"
-        component={FeedNavigator}
-      />
-      <Tab.Screen
-        name="Chat"
-        component={ChatNavigator}
-      />
-    </Tab.Navigator>
-  )
-}
 
 const TabNavigator = ({ navigation, route, theme }) => {
   const tabNavigatorProps = navigationOptions.tabNavigatorProps({ theme, route })
@@ -95,8 +64,8 @@ const TabNavigator = ({ navigation, route, theme }) => {
   return (
     <Tab.Navigator {...tabNavigatorProps}>
       <Tab.Screen
-        name="Home"
-        component={HomeNavigator}
+        name="Feed"
+        component={FeedNavigator}
         {...feedTabScreenPropsCard}
       />
       <Tab.Screen
